Move database changes to next update file
