+++ system/ee/installer/updates/ud_6_03_03.php
@@ -25,33 +25,7 @@ class Updater
      */
     public function do_update()
     {
-        $steps = new \ProgressIterator(
-            [
-                'addConditionalFieldRequiredSyncFlag',
-            ]
-        );
-
-        foreach ($steps as $k => $v) {
-            $this->$v();
-        }
-
         return true;
-    }
-
-    private function addConditionalFieldRequiredSyncFlag()
-    {
-        if (! ee()->db->field_exists('conditional_sync_required', 'channels')) {
-            ee()->smartforge->add_column(
-                'channels',
-                array(
-                    'conditional_sync_required' => array(
-                        'type' => 'CHAR(1)',
-                        'null' => false,
-                        'default' => 'n'
-                    )
-                )
-            );
-        }
     }
 }
