+++ system/ee/installer/updates/ud_6_03_04.php
@@ -1,0 +1,58 @@
+<?php
+
+/**
+ * This source file is part of the open source project
+ * ExpressionEngine (https://expressionengine.com)
+ *
+ * @link      https://expressionengine.com/
+ * @copyright Copyright (c) 2003-2022, Packet Tide, LLC (https://www.packettide.com)
+ * @license   https://expressionengine.com/license Licensed under Apache License, Version 2.0
+ */
+
+namespace ExpressionEngine\Updater\Version_6_3_4;
+
+/**
+ * Update
+ */
+class Updater
+{
+    public $version_suffix = '';
+
+    /**
+     * Do Update
+     *
+     * @return TRUE
+     */
+    public function do_update()
+    {
+        $steps = new \ProgressIterator(
+            [
+                'addConditionalFieldRequiredSyncFlag',
+            ]
+        );
+
+        foreach ($steps as $k => $v) {
+            $this->$v();
+        }
+
+        return true;
+    }
+
+    private function addConditionalFieldRequiredSyncFlag()
+    {
+        if (!ee()->db->field_exists('conditional_sync_required', 'channels')) {
+            ee()->smartforge->add_column(
+                'channels',
+                array(
+                    'conditional_sync_required' => array(
+                        'type' => 'CHAR(1)',
+                        'null' => false,
+                        'default' => 'n'
+                    )
+                )
+            );
+        }
+    }
+}
+
+// EOF
